+++ scan/scan_long.c
@@ -1,5 +1,5 @@
 #include "scan.h"
 
 size_t scan_long(const char *src,long *dest) {
-  return scan_longn(src,-1,dest);
+  return scan_longn(src,(size_t)(-1),dest);
 }
